// src/types/api.ts
export interface components {
    schemas: {
// ...
        configFullMetadata: {
// ...
            dependencies: {
                /** @description Configs referenced by this config (nested tree, max 2 levels) */
                children: components["schemas"]["configReference"][];
                /** @description Configs that reference this config (nested tree, max 2 levels) */
                parents: components["schemas"]["configReference"][];
            };
// ...
        };
// ...
    };
    responses: {
        /** @description BadRequest */
        "400BadRequest": {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["error"];
            };
        };
        /** @description Not Found - If client does not exist */
        "404NotFound": {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["error"];
            };
        };
        /** @description conflict */
        "409Conflict": {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["error"];
            };
        };
        /** @description Unprocessable Entity */
        "422UnprocessableEntity": {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["error"];
            };
        };
        /** @description Internal Server Error */
        "500InternalServerError": {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["error"];
            };
        };
    };
// ...
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    getFullConfig: {
        parameters: {
            query: {
                /** @description The id of the requested schema */
                schemaId: components["schemas"]["schemaId"];
            };
            header?: never;
            path: {
                /** @description The name of the config */
                name: components["parameters"]["ConfigNamePath"];
                version: "latest" | components["schemas"]["version"];
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["configFullMetadata"];
                };
            };
            400: components["responses"]["400BadRequest"];
            404: components["responses"]["404NotFound"];
            500: components["responses"]["500InternalServerError"];
        };
    };
// ...
    getFullSchema: {
        parameters: {
            query: {
                /** @description The id of the requested schema */
                id: components["schemas"]["schemaId"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
// ...
                        dependencies: {
                            /** @description All parent schemas that reference this schema (nested tree structure) */
                            parents: components["schemas"]["schemaReference"][];
                            /** @description All child schemas referenced by this schema (nested tree structure) */
                            children: components["schemas"]["schemaReference"][];
                        };
// ...
                    };
                };
            };
            400: components["responses"]["400BadRequest"];
            404: components["responses"]["404NotFound"];
            500: components["responses"]["500InternalServerError"];
        };
    };
// ...
}